[Behat] Extract Customer transformer
